requirements.txt
rich_click
rich
frontmatter
pyyaml
gql[all]
pydantic